feat: expose transform response via lib
